+++ lib/unifig/providers/local.rb
@@ -8,14 +8,9 @@ module Unifig
         :local
       end
 
-      # @private
-      def self.load(vars)
-        @vars = vars
-      end
-
       def self.retrieve(var_names)
         var_names.to_h do |name|
-          [name, @vars[name].local_value]
+          [name, Vars[name].local_value]
         end
       end
     end